+++ CHANGELOG.md
@@ -2,13 +2,20 @@
 
 - add reload button if necessary - for real time data
 
+<<<<<<< HEAD
 # 2.3.8
 
 - add system notifications
+=======
+# 2.4.0
+
+- add panel notifications setup (hra-done)
+>>>>>>> main
 
 # 2.3.7
 
 - add salons change direction mode (hra-done)
+- add new order invoice print (hra-done)
 
 ## Features